Adding jsdocs
